+++ lib/src/home/view/screens/initial_disclosure_agreement.dart
@@ -25,16 +25,36 @@ class _InitialDisclosureController extends WebPageController {
 
   @override
   List<Widget> children05(BuildContext context, [WebPage? widget]) {
+    //
     final _screenSize = MyApp.screenSize;
+
     final _smallScreen = MyApp.inSmallScreen;
+
     // Sometimes this method is called before the widget is mounted.
     if (_widget == null && widget != null && widget is InitialDisclosure) {
       _widget = widget;
     }
+
+    // Image of the disclosure
     final _disclosure = Image.asset(
       'assets/images/disclosure.png',
       fit: BoxFit.cover,
     );
+
+    Widget child;
+
+    // Supply interaction on the mobile phone.
+    if (_smallScreen) {
+      child = InteractiveViewer(
+        maxScale: 3,
+        minScale: 1,
+        child: _disclosure,
+      );
+    } else {
+      child = _disclosure;
+    }
+    const text =
+        "We'll sign a Non-Disclosure agreement so you can share your idea with me.";
     return [
       if (_widget?.banner ?? true)
         SizedBox(
@@ -45,38 +65,46 @@ class _InitialDisclosureController extends WebPageController {
             fit: BoxFit.cover,
           ),
         ),
-      if (_smallScreen)
-        InkWell(
-          splashColor: Colors.transparent,
-          hoverColor: Colors.transparent,
-          onTap: () => PopupPage.window<void>(
-            context,
-            (_) => Center(
-              child: Padding(
-                padding: const EdgeInsets.all(0),
-                child: InteractiveViewer(
-                  maxScale: 3,
-                  minScale: 1,
-                  child: _disclosure,
-                ),
-              ),
+      Text(
+        'Got an Idea for an App?',
+        style: TextStyle(fontSize: _smallScreen ? 24 : 48),
+      ),
+      if (_smallScreen) const AutoSizeText(text) else const Text(text),
+      InkWell(
+        splashColor: Colors.transparent,
+        hoverColor: Colors.transparent,
+        onTap: () => PopupPage.window<void>(
+          context,
+          (_) => Center(
+            child: Padding(
+              padding: const EdgeInsets.fromLTRB(200, 100, 200, 50),
+              child: child,
             ),
           ),
-          child: Padding(
-            padding: const EdgeInsets.all(0),
-            child: _disclosure,
+        ),
+        // child: Padding(
+        //   padding: const EdgeInsets.fromLTRB(300, 200, 300, 10),
+        //   child: _disclosure,
+        // ),
+        child: Container(
+          margin: EdgeInsets.fromLTRB(
+            _screenSize.width * (_smallScreen ? 0 : 0.2),
+            _screenSize.height * (_smallScreen ? 0.1 : 0.1),
+            _screenSize.width * (_smallScreen ? 0.0 : 0.2),
+            _screenSize.height * (_smallScreen ? 0.1 : 0.05),
           ),
+          child: _disclosure,
         ),
-      if (!_smallScreen)
-        Iframe(
-          height: '1000',
-          width: '640',
-          src:
-              'https://docs.google.com/document/d/e/2PACX-1vSUGrouFhcTcJqJ4ItmKU0qY6XyF4DIPf619wrlEOgvj-kubE9gV3ME_8K-o04vZqHm4nVEUQcSHVSc/pub?embedded=true',
-        ),
+      ),
+      // if (!_smallScreen)
+      //   Iframe(
+      //     height: '1000',
+      //     width: '640',
+      //     src:
+      //         'https://docs.google.com/document/d/e/2PACX-1vSUGrouFhcTcJqJ4ItmKU0qY6XyF4DIPf619wrlEOgvj-kubE9gV3ME_8K-o04vZqHm4nVEUQcSHVSc/pub?embedded=true',
+      //   ),
       const Hyperlink(Text('Print'),
           'https://docs.google.com/document/d/1Z2zdFIv_n1HXT2rkU8n6kAOr144b6nHsGbhkUOO9IA0/edit?usp=sharing')
-//      DestinationHeading(screenSize: _screenSize),
     ];
   }
 }